Updated firmware to get a more stable WHAD device.
Had some issues while in Central mode (kept connecting even if
stopped) and some other minor issues now fixed.

+++ main/inc/ble_hack.h
@@ -274,7 +274,7 @@ struct llcp_pdu_tag
     /// List element for chaining
     struct co_list_hdr hdr;
     /// Node index
-    uint16_t idx;
+    uint32_t idx;
     /// Pointer on the pdu to send
     void *ptr;
     uint16_t pdu_length;
